third_party/WebKit/LayoutTests/platform/linux/virtual/mojo-loading/http/tests/devtools/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: line 23: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 44: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 62: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
CONSOLE ERROR: line 68: Uncaught (in promise) 42
CONSOLE ERROR: line 73: Uncaught (in promise) 1e+100
CONSOLE ERROR: line 78: Uncaught (in promise) foo
CONSOLE ERROR: line 83: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 88: Uncaught (in promise) undefined
CONSOLE ERROR: Uncaught (in promise) TypeError: Failed to register a ServiceWorker: A bad HTTP response code (404) was received when fetching the script.
Tests that uncaught promise rejections are logged into console.

console-uncaught-promise.html:23 Uncaught (in promise) Error: err1
promiseTest1 @ console-uncaught-promise.html:23
timeout @ console-uncaught-promise.html:16
Promise rejected (async)
promiseTest1 @ console-uncaught-promise.html:26
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:44 Uncaught (in promise) Error: err2
promiseTest2 @ console-uncaught-promise.html:44
timeout @ console-uncaught-promise.html:16
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:40
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:39
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:38
Promise rejected (async)
promiseTest2 @ console-uncaught-promise.html:37
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:62 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
throwDOMException @ console-uncaught-promise.html:62
catcher @ console-uncaught-promise.html:54
Promise rejected (async)
promiseTest3 @ console-uncaught-promise.html:53
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:68 Uncaught (in promise) 42
promiseTest4 @ console-uncaught-promise.html:68
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:73 Uncaught (in promise) 1e+100
promiseTest5 @ console-uncaught-promise.html:73
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:78 Uncaught (in promise) foo
promiseTest6 @ console-uncaught-promise.html:78
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:83 Uncaught (in promise) {foo: 42}
    foo: 42
    __proto__: Object
promiseTest7 @ console-uncaught-promise.html:83
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
console-uncaught-promise.html:88 Uncaught (in promise) undefined
promiseTest8 @ console-uncaught-promise.html:88
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
A bad HTTP response code (404) was received when fetching the script.
console-uncaught-promise.html:1 Uncaught (in promise) TypeError: Failed to register a ServiceWorker: A bad HTTP response code (404) was received when fetching the script.
Promise (async)
promiseTest9 @ console-uncaught-promise.html:93
timeout @ console-uncaught-promise.html:16
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:15
(anonymous) @ VM:1
